printing: Update PPDResolver to map Zebra printers to generic PPD file
Zebra label printers often share the same generic PPD files
supplied by their respective manufacturers. Currently, to map each of
these printers to the correct PPD file, we need to update the PPDIndex
manually to include the usb ids of the specific printers.

With this we can map all printers that support the generic PPD file to
the same one in the PPDIndex without manually adding the device usb
ids in the index.

Bug: b:244569959
Change-Id: I0a94dbaa0e58a0ff698c914001b95bc052d3f75f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4650706
Commit-Queue: Liam Huffman <liamhuffman@google.com>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171884}