ash: Disable a test when EnablePerDeskZOrder is enabled
This CL disables the test DesksTest.VisibleOnAllDesksGlobalZOrder when
the EnablePerDeskZOrder flag is enabled. That test is written to verify
the legacy behavior and is thus expected to fail when per-desk z-order
is enabled.

The per-desk z-order feature has its own (extensive) set of tests in
`DesksPerDeskZOrderTest`.

BUG=b:291786345

Change-Id: I022240b1742a93341ef42b3102026f3a4b545939
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4697733
Commit-Queue: Daniel Andersson <dandersson@chromium.org>
Reviewed-by: Yongshun Liu <yongshun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171939}