Augment details in GetPrinterDriverInfo for CUPS IPP
Add more details to the information provided for CUPS IPP drivers.
This should help in particular for some kinds of drivers that are added
through the CUPS admin.  For example, RAW printers would otherwise only
show up text such as with "Local RAW Printer".  This update at least
provides the possibility for a user-identified name to appear, which
while not guaranteed definitive could still help with debugging in some
cases.

Including the CUPS user agent also ensures that the version of the
printing library is captured, which might be useful when there are bugs
related to particular CUPS library versions.

Fixed: 1458058
Change-Id: Ieb43c8ae6b78f5bb3578298130095121f5a85271
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4671311
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1171967}