Add a CookieMover in the DiceSignedInProfileCreator
This is the first CL in a series of CLs which allow support for profile
for third party federated accounts. This will allow cookies of the iDp
to be moved to a profile creayed via enterprise signin interception s
that the user doesn't have to signin another time nor that the user
stays signed in the previous account.

Bug: 1444802
Change-Id: Ifc7e688a3862db93ef8791c7f0ded8c2c8a2d95a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4525116
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Yann Dago <ydago@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171977}