[gbb-cct] Activator for page insights module
PageInsightsActivator monitors the dynamic settings to instantiate
page insights component at the first time when all the conditions are
met, such as sign-in, history sync without a custom passphrase, and
sWAA. It replaces SwaaChecker (which is now also one of the settings
it manages) as an interface for UI coordinator.

Low-Coverage-Reason: main class PageInsightsActivator is fully covered.
Bug: 1448966
Change-Id: If5ff112f21f5690eb885bc2c8d5e6433edd53da1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4694464
Reviewed-by: Kevin Grosu <kgrosu@google.com>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1171995}