Make variation seed available to run smoke tests on webview.
We will need to keep app_dir when running webdriver on Android/webview
because we place the seed file in a app dir before we launch the app.

Note this CL only makes the seed available for webview shell app to consume, but the test is currently not being set up on any builder/bot to run.


Bug: 1450425
Change-Id: I4363d665d6264db9c4d75592087a8445a0f55050
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4697438
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>
Cr-Commit-Position: refs/heads/main@{#1172000}