[DIPS] Update deletion latency metric to ignore scheduling time
Previously, the deletion latency metric included the time from requesting a deletion to completing a deletion. This couples the time it takes for the task to be scheduled with the execution of the task. Some very large P99 latencies have been recorded, which may be simply due to the scheduling time on the low priority thread. Now, limit dips deletion latency metric to specify just the deletion time, ignoring the time to schedule deletion on the low priority thread. This will decouple scheduling and deletion time so that the large tail latencies can be better understood.

OBSOLETE_HISTOGRAM[Privacy.DIPS.DeletionLatency]=this metric was too broad and has been updated in Privacy.DIPS.DeletionLatency2 to be more specific to the actual deletion time without including any extra time spent scheduling the task.

Bug: 1446681
Change-Id: If16a32baf5fe3120c0f595ab677a3a40c1f74191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4678180
Reviewed-by: Joshua Hood <jdh@chromium.org>
Commit-Queue: Patricia Alfonso <trishalfonso@google.com>
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172007}