mediapipe: add standalone target
To ease the pain of having double the targets this adds a gn
template that generates the two targets. That is, the one
depending on base and the one that doesn't depend upon base.
The templates are mostly drop in replacements, with a few
differences:

. dependencies that have a _standalone variant need to be
  called out explicitly. I would have liked to automate this,
  but this didn't seem possible.
. configs/public_configs get different names because of how
  these are used. By that I mean you generally add and remove
  from configs rather than reset the configs completely.

The template also adds an assert that nothing depends on
base.

Change-Id: I9997a1003fc85b7ffcf598612b7ae0c01db562dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4697428
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172047}