Add logs and crash dumps if OPRF request fails.
We add logs and generate a non fatal crash dump if the channel is
pre-stable, and the exact net error codes are reached. This will
be used to help understand why such a large percentage of OPRF
responses (~20%), seem to be failing.

We see that a subset of OPRF requests are returning a failed response.
1. Net code = -105 // The host name could not be resolved
2. Net code = -379 // The server returned a non-2xx HTTP response code.
3. Net code = -501 // The server response was insecure.

This change will be reverted after debugging and submitting fix.

Privacy proposal:go/diagnosing-check-membership-failures
BUG=chromium:1441199, b/290111610

Change-Id: I9c723bcd87f5205679be20ae011fe69beeef7626
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4687783
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Hirthanan Subenderan <hirthanan@google.com>
Cr-Commit-Position: refs/heads/main@{#1172062}