Explicitly order hide/unhide operations for pip player controls
Video overlay used to depend on implicit behaviours of
OnNativeWidgetMove and mouse moved events. This is problematic in lacros
because the timing and frequencies are different. In lacros, it is
possible for OnNativeWidgetMove and mouse move events to be several ms
apart. This leads to an observed flickering issue. Note that the rapid
visibility state change can also be observed in all platforms, but is
fast enough that the defect isn't visible. It also implicitly
assumes that a mouse move event will always follow the last widget move
callback.

This change explicitly orders the visibility control during move. When
the widget starts moving, the controls are hidden and all future updates
to the visibility are queued until the widget stops moving. 100ms seems
to be a reasonable number where the flickering cannot be observed and
feels responsive immediately after dropping the widget.

Fixed: 1419589
Change-Id: Ia53a91eb62731f29ff3e837681f94e08e7ce7fb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4685351
Commit-Queue: Fred Shih <ffred@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172063}