[Extensions] Move the page load callbacks to a separate data structure.
This is one in a series of CLs that will move data out of the
ExtensionWebRequestEventRouter that is not keyed by BrowserContext.
This will allow for a more straightforward approach when breaking it
up into multiple per-BrowserContext instances.

This CL also adds a TODO to investigate why this particular data
structure is global, rather than per-BrowserContext.

No behavior change is expected.

Bug: 1433136
Change-Id: I524fa68e9c51a69b9f1943f05341b0ad68b06f6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4697744
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172064}