WebUI: Delete obsolete a11y tests based on js2gtest, part 1.
In this part deleting all non-CrOS such tests.

The WebUI AccessibilityTest mechanism is being removed, since it has
not proven very useful, is basically abandoned for a long time and
relies on js2gtest which WebUI tests are moving away from. Porting it
off of js2gtest does not seem like a good time investment, given that
such tests have not proven to be useful. Therefore removal seems
to be the most practical and appropriate option to unblock the removal
of js2gtest.

Bug: 1457360,1002620
Change-Id: Ica3a4c14295a31aa284105b48f1e4ae96dfc539c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4680260
Commit-Queue: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172075}