[MirrorSync] Create Lit component for Inline Status
This new component displays whether files and folders
are pinned, not pinnable, queued, in progress, or none of those.

The component isolates CSS that was previously located in the
file_manager.css and file_manager_gm3.css into a more succinct
definition that is also co-located with the rendering logic.

This CL reduces code complexity related to inline sync status
and related concepts. It also helps pave the way for efforts in
converting our current virtual list into Lit.

Note that the logic that added 300ms before dismissing
the completed sync status of fully synced files has been
temporarily removed as it was not working realiably in the
first place. It will be reintroduced in a follow up CL.

Bug: b:291173347
Test: browser_tests *FilesAppBrowserTest.Test*
Change-Id: I86463bfc875791b6e62d54e200c2d40a82172e46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4688705
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Marcello Salomao <msalomao@google.com>
Cr-Commit-Position: refs/heads/main@{#1172098}