[Extensions] Remove a data member from ExtensionWebRequestEventRouter.
This is one in a series of CLs that will move data out of the
ExtensionWebRequestEventRouter that is not keyed by BrowserContext.
This will allow for a more straightforward approach when breaking it
up into multiple per-BrowserContext instances.

There is an ExtensionWebRequestTimeTracker member of this object that
is tracking timings across BrowserContext instances. This CL moves that object into a separate global instance.

No behavior change is expected.

Bug: 1433136
Change-Id: Ic0bb07ee2b44a1782b4be27fdc536ef382a91814
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4697431
Reviewed-by: Solomon Kinard <solomonkinard@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172099}