[FPS] Remove dead same_party_status member and associated dead code
This CL removes the dead/no-op CookieAccessParams::same_party_status
member, as well as a bunch of code that becomes dead as a result. This
CL does not change any behavior, since the field that is removed is
an enum with only one variant.

There is still more dead code that can be removed (e.g. the
party_context setup, and SameParty support in CanonicalCookie), but
that will be handled in followups.

Bug: 1348588
Change-Id: I7995f7a5d74198049e2125d0105b21c458e0691b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4690406
Reviewed-by: Avi Drissman <avi@chromium.org>
Auto-Submit: Chris Fredrickson <cfredric@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172139}