Migrate CrOS functions to new MediaDeviceSaltService API
This CL migrates the CrOS camera app and webcam extension API to
use the new per-storage-key API of MediaDeviceSaltService.

Since these were the last call sites, the global-salt API is
removed. The service is internally still using the global-salt storage
backend, but all call sites are ready for the new backend.

This CL also removes the global-salt version of GetSalt() in the
service and updates fieldtrial_testing_config.json so that some tests
use the new backend.

Bug: 1410462
Change-Id: Ica7b45d781ab81f047af9a868723dc604acd4cd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4678900
Reviewed-by: Kyle Williams <kdgwill@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Wei Lee <wtlee@chromium.org>
Reviewed-by: Palak Agarwal <agpalak@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172172}