[ParkableString] Prevent compression again once it failed
Currently if compressed size is not enough for parking, and then
compressed data is not stored, same compression work can be
scheduled for the same string and repeat the failing. This causes
repeatedly alloc-free large memory in the compression work.

To prevent the problem, this CL makes don't try to compression
again once it failed.

Bug: N/A
Change-Id: I6a7a854bee4e96daff9bf2ad2bc47b826ee2c837
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4686909
Reviewed-by: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: JunHo Seo <junho0924.seo@lge.com>
Cr-Commit-Position: refs/heads/main@{#1172200}