[App Details] Fix launch button for web apps from the Play Store
Previously, there was an issue involving the launch button for web apps from the Play Store wherein onclick, the app's page in the play store would not be opened correctly due to the incorrect ID being provided to the AppManagementPageHandler for opening. This CL fixes that.

Bug: 1295875
Change-Id: Ib9a74e54bbc7132ad969e1a6edc2c1ab3db84d4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4661846
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Sharmin Zaman <sharminzaman@google.com>
Cr-Commit-Position: refs/heads/main@{#1172242}