Reenable a manual password generation test
PasswordGenerationIntegrationTest.testManualGenerationUsePassword was
disabled because it was flaky. In the meantime the layout of
keyboard accessory changed so toggleAccessorySheet method needed to be
updated. To fix the flakiness it was necessary to wait until all the
UI elements have been inflated (the view holder and the buttons inside).

Bug: 1010540
Change-Id: I02e7d4b1717baa4da3f7dc8bdd15dc269bbb7075
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4693787
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1172283}