Distinguish new users skipping profile migration.
To enable Lacros, users have to complete profile migration and have it
marked as completed. For new users who are added when Lacros is already
enabled, profile migration is marked as completed due to there not being
any data that has to be migrated.

Currently there is no way to distinguish these users from users who've
gone from Ash to Lacros while undergoing profile migration. This can
make debugging harder from user reports.

This patch makes it possible to distinguish the two groups by
introducing `MigrationMode::kSkipForNewUser`.

Note that the introduction of this mode changes the meaning of
`MigrationMode`, which previously corresponded to the type of
`MigratiorDelegate`, to something that is used primarily to mark what
type of migration was completed.

In subsequent patches, the remaining use of `MigrationMode` in the old
way, e.g. passing mode to `BrowserDataMigrator::Migrate()` will be
removed. This is part of an effort to remove copy migrator.

Bug: 1464442
Test: tast run $DUT lacros.MigrateBasic, unit_tests, browser_tests
Change-Id: I980c9661d91b7909d1312aa38e745cf86a9244ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4680951
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Yuta Hijikata <ythjkt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172304}