[ios] Refactor remaining code about UserSelectableType::kPayments
No behavioral changes: this is a follow-up to
https://crrev.com/c/4684131 that further unifies the payments case to
the general-purpose code for other sync types.

Most of the logic is shared and the only special-casing required for
kPayments (i.e. coupling with kAutofill) can be expressed very concisely
with a few lines. As per added TODO, even this special-casing should be
eventually removed.

Change-Id: Icd3ea72b34a3695627cbe446a4f6c96ab52b64fc
Bug: 1459963
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4697857
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172325}