[CDP] Check drag start mouse move is handled
It's possible that the mouse movement that starts a drag is acked
after dragging starts. When this happens,

  1. `StartDragging` will update the drag state with the mouse
    movement (through `last_mouse_move_`), then
  2. the mouse movement (this time from acking in `OnInputEventAck`)
    will attempt to update the drag state _again_.

Since we only want the mouse movement to update the drag state once,
we attempt to stop it here.

Since we don't have unique identifiers for each event, the best we
can hope for is the timestamps are the same.

Note that in general, the mouse movement will be acked before the
dragging starts, so this should happen rarely.

Bug: 1430420
Change-Id: I4f8919064878e1b6685f59abf0bfc1729d1ff724
Fixed: 1459188
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4686781
Commit-Queue: Randolf Jung <jrandolf@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172340}