Clean up WebGL version tag
Cleans up the generation and most uses of the webgl-version-X tags in
GPU code. We formally split the two suites a while back, so there is no
need to keep the tag around.

The tags are still present in the ignored tag list for now until we can
be sure that any data with the tags present has aged out.

Bug: 1140283
Change-Id: Ie9212a38f38551b4749aab6149eabf1e9d9564a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4697607
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172351}