Chrome App windows: Respect OTR WindowOpenDisposition
Chrome App windows force the opening of links in new windows to be done
in a new tab instead. This did not consider whether OFF_THE_RECORD
(incognito) was specified, so a link a user intended to open in an
incognito window would be opened in a non-incognito window.

We now no longer force the WindowOpenDisposition in this case.

Bug: 1464784
Change-Id: I5a4e5bf49a0ded846b019f01ee4952a67797470e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4698179
Reviewed-by: Trent Apted <tapted@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172358}