List all inputs to rust targets in GN and verify them during build
RBE requires all inputs to be named in order for it to copy them to
the remote build environment.

GN analyze requires all inputs to be named in order to decide what
needs to be rebuilt based on a list of modified files. The current GN
analyze setup was supposed to do this as well, but it is not working.

Have gnrt produce a list of all (possible) input files for 3p and
stdlib crates so that it can be included in the generated BUILD.gn
files.

Then verify in rustc_wrapper.py that all inputs rustc reports are
found in the GN rule's sources or inputs fields by comparing with the
depfile produced from rustc.

Bug: 1459444
Change-Id: I74ce3bfa9a6404be3d3493c1bd6fc0f790809667
Cq-Include-Trybots: luci.chromium.try:android-rust-arm32-rel,android-rust-arm64-dbg,android-rust-arm64-rel,linux-rust-x64-dbg,linux-rust-x64-rel,mac-rust-x64-dbg,win-rust-x64-rel,win-rust-x64-dbg
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4673708
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: danakj <danakj@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Reviewed-by: Zequan Wu <zequanwu@google.com>
Cr-Commit-Position: refs/heads/main@{#1172375}