[Multi Apps] Disable permission toggles on sub app settings page
This change adds an explicit `disabled` property to
`AppManagementToggleRowElement`. This is set to false by default. The
enabled/disabled state of the toggle in the toggle row is now determined
as `(disabled || managed)`. `AppManagementPermissionItemElement` sets
the `disabled` property of its toggle row to (managed || is-sub-app).

Bug: 1464567
Change-Id: Ie476f0cabdb146a90daf9a8e2fdb4cb38155f7c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4678740
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Sam Thiesen <samicolon@google.com>
Cr-Commit-Position: refs/heads/main@{#1172438}