[Tab Group Dialog] Use visibility listener for phones
When hiding the TabGridDialog in the TabSwitcher context on phones it
currently appears as though the stand-in tab bitmap is purged too early.
This leaves an empty space which looks bad at reduced animation speed.
Fix this by copying Tablets and use the VISIBILITY_LISTENER to remove
the bitmap only after the animation is finished.

There is added difficulty in this because the VISIBILITY_LISTENER is
bound to both the TabGroupUi and TabSwitcher versions of the dialog and
there are two separate TabGridDialog{Mediator/Coordinators} and property
models, but only one TabGridDialogView.

Because TabGridDialogView is a singleton, but is controlled by two
separate models there is a significant opportunity for bugs. Unlike the other properties, the correct VISIBILITY_LISTENER is not the same between the two models. To fix this issue a new BINDING_TOKEN property is added which corresponds to the hashCode() of the current
TabGridDialogMediator using the DialogView.

BINDING_TOKEN
* == null - a TabGridDialogMediator may take ownership by setting itself
* new int - a TabGridDialogMediator is taking ownership. It is illegal
  state if another TabGridDialogMediator already has ownership. When
  installing a new int all the TabGridPanelProperties state is re-bound
  using the Model for the mediator taking ownership
  (guards were added to cases where this may pose an issue).
* existing int - no-op or process the key that is already there.

Demo at 10x speed reduction:
https://drive.google.com/file/d/1Un13pD2-i4F6UGNwo3Ffexy51z2Cpasy/view?usp=sharing

See bug for the "before" case.

Bug: 1464665
Change-Id: I912241904e33f12c1319a202e5ac360be92f9e46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4684992
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Mei Liang <meiliang@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172442}