[Nearby Presence] Add mojo piping to get local credentials from NP lib
Adds mojo functionality to fetch the local device's shared credentials
from the NP library. This will be used in a follow up CL to get the
local device's shared credentials and upload them to the server for
daily syncs. See
go/cros-nearby-presence-credential-web-integration

At a high level, the design doc discusses downloading remote devices'
credentials from the Nearby Presence server, and plumbing them to the
Nearby Presence library, which lives in a sandboxed utility process, as
well as retrieving the local device's credentials from the Nearby
Presence library, and uploading them to the Nearby Presence server.
These credentials will be used to authenticate devices that are
advertising nearby. This CL is one part of this: it adds the mojo
piping to retrieve the local device's credentials from the Nearby
Presence library, and in a follow up CL, it will be uploaded to the
Nearby Presence library.

This CL depends on fakes added in cl/541023409

Tested: unit tests only
Bug: b/276307539
Change-Id: Ie9527fef6444e744a55e493fe14c5bf5f3a05865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4622767
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Juliet Lévesque <julietlevesque@google.com>
Cr-Commit-Position: refs/heads/main@{#1172459}